bug fixed

+++ js/modalHendler.js
@@ -1,101 +1,49 @@
 function menuEventsSetter(marker) {
-    let catchTheMenu = document.getElementsByClassName("new-modal");
+    let catchTheMenu =  $(".new-modal");
     console.log(catchTheMenu.length);
 
-    $(".new-modal").unbind().click((e) => {
-        // document.getElementsByClassName("new-modal")[0].addEventListener("click", function (e) {
-        // $('.new-modal').one("click", function (e) {
+    $(".new-modal")[0].onclick = (e)=>{
+    // $(".new-modal").off('click').on('click', (e)=>{
+    // document.getElementsByClassName("new-modal")[0].addEventListener("click", function (e) {
+        // $('.new-modal').on("click", function (e) {
         if (e.target.id == 'close') {
-            $(this).hide();
+            $(".new-modal").hide();
             $('.icon-first_menu').empty();
         } else {
 
             let check_modal = $('.new-modal')[0].id;
-
             if (check_modal == 'third_menu_modal' || check_modal == 'second_menu_modal') {
                 if (e.target && e.target.src) {
                     icon = e.target.src;
                     if (e.target.id == 'b1') {
-                        openModal(e.target.id);
-                        return;
+                        openModal(e.target.id, marker);
                     } else {
-
                         app.setIcon(icon);
-
-                        setMarkerIcon();
-
+                        setMarkerIcon(icon, marker);
+                        objectForSave = e.target;
                         app.setCount();
                         console.log('count', app.getCount());
-
                         return;
                     }
+                    // saveDB(objectForSave, marker);
                 }
-            } else {
-                openModal(e.target.id);
-                // else if (app.getA() && app.getIcon()) return;
-            }
+            } else if (check_modal == 'first_menu_modal') openModal(e.target.id, marker);
+            // else if (app.getA() && app.getIcon()) return;
         }
-    });
+    };
+    // setTimeout(() => {
+    //     //wait for js
+    // }, 100);
 }
 
-function setMarkerIcon() {
+function setMarkerIcon(icon, marker) {
     let new_marker_icon = app.getIcon();
 
     if (new_marker_icon) {
-        let mymap = app.getGoogleMap();
-        let LeafIcon = app.getLeafMarker();
-        let latlng = app.getNewLocation();
-        let new_icon;
-
         app.setIcon(new_marker_icon + '/revision/latest/scale-to-width-down/' + sizeMap(new_marker_icon));
-
-        console.log('getLocations', getLocations);
-
-        if (!LeafIcon) {
-
-            // let mymap = app.getGoogleMap();
-            let LeafIcon = L.Icon.extend({
-                options: {
-                    shadowUrl: app.getIcon(),
-                    iconSize: [120, 120],
-                    shadowSize: [30, 34],
-                    iconAnchor: [22, 34],
-                    shadowAnchor: [4, 32],
-                    popupAnchor: [-3, -36]
-                }
-            });
-
-            new_icon = new LeafIcon({
-                iconUrl: app.getIcon()
-            });
-        } else {
-
-            new_icon = new LeafIcon({
-                iconUrl: app.getIcon()
-            });
-        }
-        // marker.setIcon(app.getIcon());
-
-        console.log(latlng.hasOwnProperty('lng'));
-
-        if (latlng.hasOwnProperty('lng')) {
-            L.marker([latlng.lat, latlng.lng], {
-                icon: new_icon
-            }).addTo(mymap).addTo(mymap).on('click', (e) => {
-                // app.setNewLocation(e.latlng);
-                open_login_edit_modal(null);
-            });
-        } else {
-            L.marker(latlng, {
-                icon: new_icon
-            }).addTo(mymap).on('click', (e) => {
-                // app.setNewLocation(e.latlng);
-                open_login_edit_modal(null);
-            });
-        }
-
+        marker.setIcon(app.getIcon());
         let new_icon_to_save = app.getIcon();
-        updateDBicons(new_icon_to_save);
+        updateDBicons(marker, new_icon_to_save);
         saveDB();
     }
 
@@ -117,7 +65,7 @@ function missionFormater(item) {
     return false;
 }
 
-function openModal(id, location) {
+function openModal(id, marker) {
     $('.lds-ripple').hide();
     let menuItems = db.main_menu;
     let subMenuItems = db.sub_menu;
@@ -136,7 +84,7 @@ function openModal(id, location) {
             console.log('second_menu_modal');
 
             buildMenu("second_menu_modal", bgColorClass2, subMenuItems || menuSearcher());
-            menuEventsSetter();
+            menuEventsSetter(marker);
 
         } else if (id.includes("b")) {
             if (id === 'b1') {
@@ -151,9 +99,10 @@ function openModal(id, location) {
             }
         }
     } else {
-        buildMenu("first_menu_modal", bgColorClass1, menuItems);
-        console.log('first_menu_modal');
-        menuEventsSetter();
+       
+            buildMenu("first_menu_modal", bgColorClass1, menuItems);
+            console.log('first_menu_modal');
+            menuEventsSetter(marker);
     }
 }
 
@@ -248,10 +197,7 @@ function buildMenu(id, bgColorClass, menuItems) {
     $(".new-modal").show();
 }
 
-function open_login_edit_modal(type) {
-    let location = app.getNewLocation();
-    let marker = app.getMarker();
-    let marker_label = app.getLabelMarker();
+function open_login_edit_modal(type, marker) {
 
     let header = "Edit Mission?",
         adminLabel = "",
@@ -259,7 +205,7 @@ function open_login_edit_modal(type) {
     adminIcon = '<i class="fa fa-cogs" id="adminMenu" aria-hidden="true" style="position: absolute;"></i>';
     //https://www.google.com/maps/search/?api=1&query=31.785492733328045, 35.214104199213125
     googleMapLink = `<div class="center" style="padding:0">
-                        <a href="https://www.google.com/maps/search/?api=1&query=${location}" target="_blank">
+                        <a href="https://www.google.com/maps/search/?api=1&query=${marker.position}" target="_blank">
                         <h6 style="position: absolute; left: calc(50% - 53px);">Get Directions</h6></a></div>`;
 
 
@@ -305,42 +251,44 @@ function open_login_edit_modal(type) {
     $(".new-modalL").show();
     $("#passwordL" + adminLabel).focus();
 
+    cBtn = document.getElementById("cancelBtnL" + adminLabel);
+    oBtn = document.getElementById("okBtnL" + adminLabel);
 
-    $("#cancelBtnL" + adminLabel).unbind().click(() => {
+    $("#okBtnL" + adminLabel).off('click').on('click',()=>{
+    // cBtn.addEventListener("click", function () {
         deleteAndHideElement($("#login_edit_modal" + adminLabel), 400);
     });
-    $("#okBtnL" + adminLabel).unbind().click(() => {
+    $("#okBtnL" + adminLabel).off('click').on('click', function(){
+    // oBtn.addEventListener("click", function () {
         if (adminLabel) {
             let tokenLadminLabel = $("#passwordL" + adminLabel).val();
 
-            // TODO: delete mardkr location.setMap(null);
+            if (tokenLadminLabel == app.getSuperToken()) {
 
-            let marker_to_delete = marker_label;
-            let mymap = app.getGoogleMap();
-            mymap.removeLayer(marker);
+                marker.setMap(null);
+                deleteMissonDB(marker.labelContent);
+                deleteAndHideElement($("#login_edit_modal" + adminLabel), 400);
 
-            deleteMissionDB(marker_to_delete);
-            deleteAndHideElement($("#login_edit_modal" + adminLabel), 400);
-
-            // delete icon
-            for (var i = getLocations.length - 1; i >= 0; i--) {
-                if (getLocations[i].label == marker_to_delete) {
-                    getLocations.splice(i, 1);
-                    break;
+                // delete icon
+                for (var i = getLocations.length - 1; i >= 0; i--) {
+                    if (getLocations[i].label == marker.labelContent) {
+                        getLocations.splice(i, 1);
+                        break;
+                    }
                 }
             }
-            // }
         } else {
             let tokenL = $("#passwordL" + adminLabel).val();
             if (tokenL == app.getToken()) {
                 deleteAndHideElement($("#login_edit_modal" + adminLabel), 400);
-                openModal(null, location);
+                openModal(null, marker);
             }
         }
     });
 
-    document.getElementById("adminMenu").addEventListener("click", function () {
+    $("#adminMenu").off('click').on('click', ()=>{
+    // document.getElementById("adminMenu").addEventListener("click", function () {
         deleteAndHideElement($("#login_edit_modal" + adminLabel), 0);
-        open_login_edit_modal("admin");
+        open_login_edit_modal("admin", marker);
     });
 }
